src/main/java/br/edu/fesa/gerenciador_gado/Models/UsuarioModel.java
/*
 * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
 * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
 */
package br.edu.fesa.gerenciador_gado.Models;

/**
 *
 * @author Lohan
 * @author Rodrigo Puertas
 */
public class UsuarioModel {
    private Integer id;
    private String nome;
    private String email;
    private String registro;
    private PerfilUsuarioModel perfilUsuario;

    public Integer getId() {
        return id;
    }

    public void setId(Integer id) {
        this.id = id;
    }

    public String getNome() {
        return nome;
    }

    public void setNome(String nome) {
        this.nome = nome;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public String getRegistro() {
        return registro;
    }

    public void setRegistro(String registro) {
        this.registro = registro;
    }

    public PerfilUsuarioModel getPerfilUsuario() {
        return perfilUsuario;
    }

    public void setPerfilUsuario(PerfilUsuarioModel perfilUsuario) {
        this.perfilUsuario = perfilUsuario;
    }
}
